Move the `Ascii85Stream` from `src/core/stream.js` and into its own file
